update curriculum check

+++ admin/update_curriculum_action.php
@@ -13,10 +13,15 @@ $choice='';
 
 include('scripts/sub_action.php');
 
-include('scripts/answer_action.php');
+$coursecheck=$_POST['coursecheck0'];
+$asscheck=$_POST['asscheck0'];
+$groupcheck=$_POST['groupcheck0'];
 
 /*reads the array $curriculum listing those required for this school-site*/
 require('../curriculum/include.php');
+
+$d_uid=mysql_query("SELECT uid FROM users WHERE username='administrator'");	
+$adminuid=mysql_result($d_uid,0);
 
 function read_curriculum_file($filename,$curriculum){
 	$path='../curriculum/'.$curriculum.'/'.$filename;
@@ -27,16 +32,16 @@ function read_curriculum_file($filename,$curriculum){
 	return $xmlArray;
 	}
 
-mysql_query("DELETE FROM cridbid");
-mysql_query("DELETE FROM classes");
-mysql_query("DELETE FROM component");
-mysql_query("DELETE FROM form");
-mysql_query("DELETE FROM yeargroup");
 
-$d_uid=mysql_query("SELECT uid FROM users WHERE username='administrator'");	
-$adminuid=mysql_result($d_uid,0);
-
-//$curriculums[]='common';
+if($coursecheck=='yes'){
+	mysql_query("DELETE FROM cridbid");
+	mysql_query("DELETE FROM classes");
+	mysql_query("DELETE FROM component");
+	}
+if($groupcheck=='yes'){
+	mysql_query("DELETE FROM form");
+	mysql_query("DELETE FROM yeargroup");
+	}
 
 while(list($index,$curriculum)=each($curriculums)){
 
@@ -44,112 +49,117 @@ while(list($index,$curriculum)=each($curriculums)){
 
 	$Courses=xmlarray_indexed_check($Courses,'course');
 	while(list($index,$Course)=each($Courses['course']) and $Course['id']!=''){
-		/*****************Courses************************/
-   		$crid=$Course['id'];
+	  /*****************Courses************************/
+   	  $crid=$Course['id'];
+
+	  if($coursecheck=='yes'){
 		if($crid!='%'){
-   		$name=$Course['name'];
-   		$sequence=$Course['sequence'];
-   		$endmonth=$Course['endmonth'];
-		$course_generate=$Course['setting'];
-   		$course_many=$Course['classes'];
-		if(is_array($Course['naming'])){
-			$course_naming=$Course['naming']['root'] 
+			$name=$Course['name'];
+			$sequence=$Course['sequence'];
+			$endmonth=$Course['endmonth'];
+			$course_generate=$Course['setting'];
+			$course_many=$Course['classes'];
+			if(is_array($Course['naming'])){
+				$course_naming=$Course['naming']['root'] 
 					.';'.$Course['naming']['stem'] 
 					.';'.$Course['naming']['branch'].';'.$Course['naming']['counter'];
-			if(sizeof($course_naming)>39){$course_naming='';}
-			}
-   		else{$course_naming='';}
+				if(sizeof($course_naming)>39){$course_naming='';}
+				}
+			else{$course_naming='';}
 
-		$d_course=mysql_query("SELECT name FROM course WHERE id='$crid'");
-		if(mysql_num_rows($d_course)==0){
-			mysql_query("INSERT INTO course (id, name, sequence,
+			$d_course=mysql_query("SELECT name FROM course WHERE id='$crid'");
+			if(mysql_num_rows($d_course)==0){
+				mysql_query("INSERT INTO course (id, name, sequence,
 				generate, naming, many, endmonth)
 				VALUES ('$crid','$name','$sequence', 
 				'$course_generate','$course_naming','$course_many','$endmonth')");
-			mysql_query("INSERT INTO groups (subject_id,
-					course_id, name) VALUES ('%', '$crid', '$crid')");
-			$gid=mysql_insert_id();
-			mysql_query("INSERT INTO perms (uid, gid, r, w, x) 
-					VALUES('$adminuid','$gid', '1', '1', '1')");
-			}
-		else{
-			mysql_query("UPDATE course SET name='$name',
-					sequence='$sequence', generate='$course_generate', naming='$course_naming',
-					many='$course_many', endmonth='$endmonth' WHERE id='$crid'");
-			}
-	
-		$Subjects=xmlarray_indexed_check($Course['subjects'],'subject');
-		while(list($index,$Subject)=each($Subjects['subject']) and $Subject['id']!=''){
-			/****Subjects*****/
-			$bid=$Subject['id'];
-			$name=$Subject['name'];
-			if(isset($Subject['setting'])){$generate=$Subject['setting'];}
-			else{$generate=$course_generate;}
-			if(isset($Subject['classes'])){$many=$Subject['classes'];}
-			else{$many=$course_many;}
-			if(is_array($Subject['naming'])){
-				$naming=$Subject['naming']['root'] 
-						.';'.$Subject['naming']['stem'] 
-						.';'.$Subject['naming']['branch'].';'.$Subject['naming']['counter'];
-				if(sizeof($naming)>39){$naming='';}
-				}
-	  		else{$naming=$course_naming;}
-			$d_subject=mysql_query("SELECT name FROM subject WHERE id='$bid'");
-			if(mysql_num_rows($d_subject)==0){
-			   mysql_query("INSERT INTO subject (id, name)
-						VALUES('$bid','$name')");
 				mysql_query("INSERT INTO groups (subject_id,
-					course_id, name) VALUES ('$bid', '%', '$bid')");
+					course_id, name) VALUES ('%', '$crid', '$crid')");
 				$gid=mysql_insert_id();
 				mysql_query("INSERT INTO perms (uid, gid, r, w, x) 
 					VALUES('$adminuid','$gid', '1', '1', '1')");
 				}
 			else{
-				mysql_query("UPDATE subject SET name='$name'
-					WHERE id='$bid'");
+				mysql_query("UPDATE course SET name='$name',
+					sequence='$sequence', generate='$course_generate', naming='$course_naming',
+					many='$course_many', endmonth='$endmonth' WHERE id='$crid'");
 				}
-
-			mysql_query("INSERT INTO cridbid SET course_id='$crid', subject_id='$bid'");
-
-			$Stages=xmlarray_indexed_check($Course['stages'],'stage');
-			while(list($index,$stage)=each($Stages['stage']) and $stage!=''){
-				mysql_query("INSERT INTO classes (many,
-								generate, naming, stage, subject_id, course_id) VALUES
-								('$many', '$generate', '$naming', '$stage', '$bid',
-								'$crid')");
-				$cohid=updateCohort(array('course_id'=>$crid,'stage'=>$stage));
-				}
-
-			$Components=xmlarray_indexed_check($Subject['components'],'component');
-			while(list($index,$Component)=each($Components['component'])
-																and $Component['id']!=''){
-				/****Components****/
-				$pid=$Component['id'];
-				$status=$Component['status'];
-				$name=$Component['name'];
+	
+			$Subjects=xmlarray_indexed_check($Course['subjects'],'subject');
+			while(list($index,$Subject)=each($Subjects['subject']) and $Subject['id']!=''){
+				/****Subjects*****/
+				$bid=$Subject['id'];
+				$name=$Subject['name'];
+				if(isset($Subject['setting'])){$generate=$Subject['setting'];}
+				else{$generate=$course_generate;}
+				if(isset($Subject['classes'])){$many=$Subject['classes'];}
+				else{$many=$course_many;}
+				if(is_array($Subject['naming'])){
+					$naming=$Subject['naming']['root'] 
+						.';'.$Subject['naming']['stem'] 
+						.';'.$Subject['naming']['branch'].';'.$Subject['naming']['counter'];
+					if(sizeof($naming)>39){$naming='';}
+					}
+				else{$naming=$course_naming;}
 				$d_subject=mysql_query("SELECT name FROM subject WHERE id='$bid'");
 				if(mysql_num_rows($d_subject)==0){
 					mysql_query("INSERT INTO subject (id, name)
-						VALUES('$pid','$name')");
-					}
-				$d_component=mysql_query("SELECT status FROM component
-					WHERE id='$pid' AND course_id='$crid' AND subject_id='$bid'");
-				if(mysql_num_rows($d_component)==0){
-					mysql_query("INSERT INTO component (id, course_id, subject_id, status)
-							VALUES('$pid','$crid','$bid','$status')");
+						VALUES('$bid','$name')");
+					mysql_query("INSERT INTO groups (subject_id,
+					course_id, name) VALUES ('$bid', '%', '$bid')");
+					$gid=mysql_insert_id();
+					mysql_query("INSERT INTO perms (uid, gid, r, w, x) 
+					VALUES('$adminuid','$gid', '1', '1', '1')");
 					}
 				else{
-					mysql_query("UPDATE component SET status='$status'
-					WHERE id='$pid' AND course_id='$crid' AND subject_id='$bid'");
+					mysql_query("UPDATE subject SET name='$name'
+									WHERE id='$bid'");
+					}
+
+				mysql_query("INSERT INTO cridbid SET course_id='$crid', subject_id='$bid'");
+
+				$Stages=xmlarray_indexed_check($Course['stages'],'stage');
+				while(list($index,$stage)=each($Stages['stage']) and $stage!=''){
+					mysql_query("INSERT INTO classes (many,
+								generate, naming, stage, subject_id, course_id) VALUES
+								('$many', '$generate', '$naming', '$stage', '$bid',
+								'$crid')");
+					$cohid=updateCohort(array('course_id'=>$crid,'stage'=>$stage));
+					}
+
+				$Components=xmlarray_indexed_check($Subject['components'],'component');
+				while(list($index,$Component)=each($Components['component'])
+																and $Component['id']!=''){
+					/****Components****/
+					$pid=$Component['id'];
+					$status=$Component['status'];
+					$name=$Component['name'];
+					$d_subject=mysql_query("SELECT name FROM subject WHERE id='$bid'");
+					if(mysql_num_rows($d_subject)==0){
+						mysql_query("INSERT INTO subject (id, name)
+								VALUES('$pid','$name')");
+						}
+					$d_component=mysql_query("SELECT status FROM component
+						WHERE id='$pid' AND course_id='$crid' AND subject_id='$bid'");
+					if(mysql_num_rows($d_component)==0){
+						mysql_query("INSERT INTO component (id, course_id, subject_id, status)
+							VALUES('$pid','$crid','$bid','$status')");
+						}
+					else{
+						mysql_query("UPDATE component SET status='$status'
+							WHERE id='$pid' AND course_id='$crid' AND subject_id='$bid'");
+						}
 					}
 				}
 			}
-			}
+		}
 
-		$AssessmentMethods=$Course['assessmentmethods'];
 
-		$GradeSchemes=xmlarray_indexed_check($AssessmentMethods['gradeschemes'],'gradescheme');
-		while(list($index,$GradeScheme)=each($GradeSchemes['gradescheme']) 
+	  if($asscheck=='yes'){
+
+ 		  $AssessmentMethods=$Course['assessmentmethods'];
+		  $GradeSchemes=xmlarray_indexed_check($AssessmentMethods['gradeschemes'],'gradescheme');
+		  while(list($index,$GradeScheme)=each($GradeSchemes['gradescheme']) 
 												and $GradeScheme['name']!=''){
 			/*****************Grade Schemes***************/
 			$name=$GradeScheme['name'];
@@ -178,8 +188,8 @@ while(list($index,$curriculum)=each($curriculums)){
 				}
 			}
 
-		$Categories=xmlarray_indexed_check($AssessmentMethods['categories'],'category');	
-		while(list($index,$Category)=each($Categories['category']) and
+		  $Categories=xmlarray_indexed_check($AssessmentMethods['categories'],'category');	
+		  while(list($index,$Category)=each($Categories['category']) and
 																   $Category['name']!=''){
 			/*****************Categories***************/
 			$name=$Category['name'];
@@ -205,8 +215,8 @@ while(list($index,$curriculum)=each($curriculums)){
 				}
 			}
 
-		$Ratings=xmlarray_indexed_check($AssessmentMethods['ratings'],'rating');
-		while(list($index,$Rating)=each($Ratings['rating']) and $Rating['name']!=''){
+		  $Ratings=xmlarray_indexed_check($AssessmentMethods['ratings'],'rating');
+		  while(list($index,$Rating)=each($Ratings['rating']) and $Rating['name']!=''){
 			/*****************Ratings***************/
 			$name=$Rating['name'];
 			$Values=xmlarray_indexed_check($Rating['values'],'value');
@@ -229,8 +239,8 @@ while(list($index,$curriculum)=each($curriculums)){
 				}
 			}
 
-		$MarkDefinitions=xmlarray_indexed_check($AssessmentMethods['markdefinitions'],'mark');
-		while(list($index,$Mark)=each($MarkDefinitions['mark']) and $Mark['name']!=''){
+		  $MarkDefinitions=xmlarray_indexed_check($AssessmentMethods['markdefinitions'],'mark');
+		  while(list($index,$Mark)=each($MarkDefinitions['mark']) and $Mark['name']!=''){
 			/*****************Mark Definitions***************/
 			$name=$Mark['name'];
 			$scoretype=$Mark['scoretype'];
@@ -256,8 +266,8 @@ while(list($index,$curriculum)=each($curriculums)){
 				}
 			}
 
-		$Methods=xmlarray_indexed_check($AssessmentMethods['methods'],'method');	
-		while(list($index,$Method)=each($Methods['method']) and $Method['value']!=''){
+		  $Methods=xmlarray_indexed_check($AssessmentMethods['methods'],'method');	
+		  while(list($index,$Method)=each($Methods['method']) and $Method['value']!=''){
 			/*****************Methods***************/
 			$subtype=$Method['value'];
 			$name=$Method['description'];
@@ -278,8 +288,8 @@ while(list($index,$curriculum)=each($curriculums)){
 				}
 			}
 
-		$ResultQualifiers=xmlarray_indexed_check($AssessmentMethods['resultqualifiers'],'resultqualifier');	
-		while(list($index,$ResultQualifier)=each($ResultQualifiers['resultqualifier'])
+		  $ResultQualifiers=xmlarray_indexed_check($AssessmentMethods['resultqualifiers'],'resultqualifier');	
+		  while(list($index,$ResultQualifier)=each($ResultQualifiers['resultqualifier'])
 														   	and $ResultQualifier['value']!=''){
 			/*****************ResultQualifierss***************/
 			$subtype=$ResultQualifier['value'];
@@ -295,12 +305,17 @@ while(list($index,$curriculum)=each($curriculums)){
 				WHERE subtype='$subtype' AND course_id='$crid' AND type='rsq'");
 				}
 			}
+		  }
 		}
 
-	$Groups=read_curriculum_file('groups.xml',$curriculum);
 
-	$YearGroups=xmlarray_indexed_check($Groups['yeargroups'],'yeargroup');
-	while(list($index,$Group)=each($YearGroups['yeargroup']) and $Group['id']!=''){
+
+
+	if($groupcheck=='yes'){
+	  $Groups=read_curriculum_file('groups.xml',$curriculum);
+
+	  $YearGroups=xmlarray_indexed_check($Groups['yeargroups'],'yeargroup');
+	  while(list($index,$Group)=each($YearGroups['yeargroup']) and $Group['id']!=''){
 		/*****************Yeargroups************************/
  		$yid=(int)$Group['id'];
    		$name=$Group['name'];
@@ -343,6 +358,7 @@ while(list($index,$curriculum)=each($curriculums)){
 			$comid=updateCommunity(array('name'=>$fid,'type'=>'form'));
 			}
 		}
+	  }
 
 	$result[]=get_string('updatedcurriculum',$book).$curriculum;
 	}
